+++ src/QueryFramework.Abstractions.Tests/GlobalUsings.cs
@@ -1,17 +1,13 @@
 ﻿global using System;
 global using System.Collections.Generic;
 global using System.ComponentModel.DataAnnotations;
-global using System.IO;
 global using System.Linq;
 global using CrossCutting.Common;
 global using CrossCutting.Common.Extensions;
 global using ExpressionFramework.Domain.Builders.Expressions;
 global using ExpressionFramework.Domain.Evaluatables;
 global using FluentAssertions;
-global using ModelFramework.CodeGeneration.CodeGenerationProviders;
 global using Moq;
 global using QueryFramework.Abstractions.Extensions;
 global using QueryFramework.Abstractions.Queries;
-global using QueryFramework.Abstractions.Tests.CodeGenerationProviders;
-global using TextTemplateTransformationFramework.Runtime.CodeGeneration;
 global using Xunit;