+++ src/QueryFramework.Abstractions/Domains/QuerySortOrderDirection.cs
@@ -1,0 +1,7 @@
+﻿namespace QueryFramework.Abstractions.Domains;
+
+public enum QuerySortOrderDirection
+{
+    Ascending,
+    Descending
+}