+++ src/QueryFramework.CodeGeneration/CodeGenerationProviders/AbstractionsBuildersInterfaces.cs
@@ -1,23 +1,22 @@
-﻿namespace CodeGeneration.CodeGenerationProviders;
+﻿namespace QueryFramework.CodeGeneration.CodeGenerationProviders;
 
 [ExcludeFromCodeCoverage]
 public class AbstractionsBuildersInterfaces : QueryFrameworkCSharpClassBase, ICodeGenerationProvider
 {
-    public override string Path => "QueryFramework.Abstractions/Builders";
-    public override string DefaultFileName => "Interfaces.template.generated.cs";
+    public override string Path => $"{Constants.Namespaces.Abstractions}/Builders";
 
     public override object CreateModel()
         => GetImmutableBuilderClasses
         (
-            GetModels(),
-            "QueryFramework.Core",
-            "QueryFramework.Core.Builders"
+            GetCoreModels(),
+            Constants.Namespaces.Abstractions,
+            Constants.Namespaces.AbstractionsBuilders
         )
         .Select
         (
             x => x.ToInterfaceBuilder()
                   .WithPartial()
-                  .WithNamespace("QueryFramework.Abstractions.Builders")
+                  .WithNamespace(Constants.Namespaces.AbstractionsBuilders)
                   .WithName($"I{x.Name}")
                   .With(x => x.Methods.RemoveAll(y => y.Static))
                   .Build()