+++ src/QueryFramework.CodeGeneration/Models/IQueryParameter.cs
@@ -1,0 +1,9 @@
+﻿namespace QueryFramework.CodeGeneration.Models;
+
+public interface IQueryParameter
+{
+    [Required]
+    string Name { get; }
+
+    object Value { get;}
+}