+++ src/QueryFramework.CodeGeneration/Models/IQueryParameterValue.cs
@@ -1,0 +1,7 @@
+﻿namespace QueryFramework.CodeGeneration.Models;
+
+public interface IQueryParameterValue
+{
+    [Required]
+    string Name { get; }
+}