+++ src/QueryFramework.Core.Tests/GlobalUsings.cs
@@ -9,6 +9,7 @@ global using ExpressionFramework.Domain.Operators;
 global using FluentAssertions;
 global using QueryFramework.Abstractions;
 global using QueryFramework.Abstractions.Builders;
+global using QueryFramework.Abstractions.Domains;
 global using QueryFramework.Abstractions.Extensions;
 global using QueryFramework.Core.Builders;
 global using QueryFramework.Core.Extensions;