+++ src/QueryFramework.Core/GlobalUsings.cs
@@ -12,6 +12,7 @@ global using ExpressionFramework.Domain.Domains;
 global using ExpressionFramework.Domain.Evaluatables;
 global using QueryFramework.Abstractions;
 global using QueryFramework.Abstractions.Builders;
+global using QueryFramework.Abstractions.Domains;
 global using QueryFramework.Abstractions.Extensions;
 global using QueryFramework.Abstractions.Queries;
 global using QueryFramework.Abstractions.Queries.Builders;