+++ src/QueryFramework.InMemory/GlobalUsings.cs
@@ -9,6 +9,7 @@ global using ExpressionFramework.Domain.Expressions;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.DependencyInjection.Extensions;
 global using QueryFramework.Abstractions;
+global using QueryFramework.Abstractions.Domains;
 global using QueryFramework.Abstractions.Queries;
 global using QueryFramework.Core.Queries;
 global using QueryFramework.InMemory.Abstractions;