+++ src/QueryFramework.SqlServer/GlobalUsings.cs
@@ -16,6 +16,7 @@ global using ExpressionFramework.Domain.Expressions;
 global using ExpressionFramework.Domain.Operators;
 global using Microsoft.Extensions.DependencyInjection;
 global using QueryFramework.Abstractions;
+global using QueryFramework.Abstractions.Domains;
 global using QueryFramework.Abstractions.Extensions;
 global using QueryFramework.Abstractions.Queries;
 global using QueryFramework.SqlServer.Abstractions;